Allow for implicit blocks
Meaning that special forms that expect blocks can now understand that a single form that isn't wrapped in a block is meant to be a block. Those special forms just wrap the form in a block themselves.

For example: Instead of writing

if(true {1} {2})

you can now write

if(true 1 2)