Make each core fn available both with its unsafe and safe name
E.g., the fn string? is now exported as both string? *and* string__QUESTION_MARK__. The latter part is necessary for the repl to work, as its context is an object that holds only the unsafe name (i.e., the one with the question mark), but the transpiler turns it into the safe name and then codes for a look up of the safe name on the context object. E.g., the transpiler will see string? and convert it to string__QUESTION_MARK__, which does not exist on the context object.

This fix makes both names available on the context object.